Ignored unsigned int vs signed int warnings in .vcxproj Fix 32-bit vs 64-bit integer issue in Rule 10 example Added missing "typename" in Rule 14 example Included <string> in Rule 16 example Fixed bug noted by Radoslaw Jurga in Rule 19 example Simplify Character destructor in Rule 2 example (to match book) Switch from "terms" to "coeffs" for clarity in Rule 3 example (to match book) Fix cut + paste bug in Rule 4 example (to match book)

+++ rule10.cpp
@@ -736,7 +736,7 @@ namespace rule10
 					for (Character * spottedCharacter : record.m_spottedCharacters)
 					{
 						auto iter = find(characters.begin(), characters.end(), spottedCharacter);
-						printf(" %d", iter - characters.begin());
+						printf(" %d", static_cast<int>(iter - characters.begin()));
 					}
 					printf("\n");
 				}
